Move EditorActions and enable mark as completed
