+++ ui/src/components/index.js
@@ -10,8 +10,8 @@ export { Plate } from './editor/Plate';
 export { NoPlatesMessage } from './editor/Plate';
 export { ComponentList } from './editor/ComponentList';
 export { Tools } from './editor/Tools';
-export { EditorActions } from './editor/EditorActions';
 export { EditorV2 } from './editor/EditorV2';
 export { ContainerTabBar } from './editor/ContainerTabBar';
 export { ContainerGrid } from './editor/ContainerGrid';
 export { ActiveContainer } from './editor/ActiveContainer';
+export { PopupButton } from './ui/PopupButton';
