+++ ui/src/implementations/kapture-backend/components/EditorActions/ImportBarcodesButton/index.js
