+++ ui/src/implementations/kapture-backend/components/EditorActions/MarkAsCompletedButton/MarkAsCompletedButton.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Message, Button } from 'semantic-ui-react';
 
-import { PopupButton } from '../../../ui/PopupButton';
+import { PopupButton } from '../../../../../components';
 import styles from './MarkAsCompletedButton.module.css';
 
 export class MarkAsCompletedButton extends Component {
